+++ src/main/java/nextstep/roomescape/controller/RequestDTO/ThemeRequestDTO.java
@@ -1,8 +1,8 @@
-package nextstep.roomescape.theme.controller.dto;
+package nextstep.roomescape.controller.RequestDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import nextstep.roomescape.theme.repository.model.Theme;
+import nextstep.roomescape.repository.model.Theme;
 
 @Getter
 @AllArgsConstructor
