+++ src/main/java/nextstep/roomescape/controller/ResponseDTO/ReservationResponseDTO.java
@@ -1,11 +1,10 @@
-package nextstep.roomescape.reservation.controller.dto;
+package nextstep.roomescape.controller.ResponseDTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import nextstep.roomescape.reservation.repository.model.Reservation;
-import nextstep.roomescape.theme.repository.model.Theme;
-import nextstep.roomescape.theme.repository.model.Theme;
+import nextstep.roomescape.repository.model.Reservation;
+import nextstep.roomescape.repository.model.Theme;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
