+++ src/main/java/nextstep/roomescape/repository/ReservationRepositoryMemoryImpl.java
@@ -1,7 +1,7 @@
-package nextstep.roomescape.reservation.repository;
+package nextstep.roomescape.repository;
 
 import nextstep.roomescape.exception.NotExistEntityException;
-import nextstep.roomescape.reservation.repository.model.Reservation;
+import nextstep.roomescape.repository.model.Reservation;
 import nextstep.roomescape.exception.DuplicateEntityException;
 import org.springframework.transaction.annotation.Transactional;
