+++ src/main/java/nextstep/roomescape/service/ThemeService.java
@@ -1,8 +1,8 @@
-package nextstep.roomescape.theme.service;
+package nextstep.roomescape.service;
 
 
-import nextstep.roomescape.theme.controller.dto.ThemeRequestDTO;
-import nextstep.roomescape.theme.repository.model.Theme;
+import nextstep.roomescape.controller.RequestDTO.ThemeRequestDTO;
+import nextstep.roomescape.repository.model.Theme;
 
 import java.util.List;
