+++ src/test/java/nextstep/roomescape/reservation/ReservationControllerTest.java
@@ -3,9 +3,9 @@ package nextstep.roomescape.reservation;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.roomescape.reservation.controller.dto.ReservationRequestDTO;
-import nextstep.roomescape.theme.controller.dto.ThemeRequestDTO;
-import nextstep.roomescape.theme.repository.model.Theme;
+import nextstep.roomescape.controller.RequestDTO.ReservationRequestDTO;
+import nextstep.roomescape.controller.RequestDTO.ThemeRequestDTO;
+import nextstep.roomescape.repository.model.Theme;
 import org.junit.jupiter.api.*;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
